llm-service/requirements.txt
llama-cpp-python[server]
fastapi
uvicorn